+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aave/cli
 
+## 4.1.0-next.9
+
+### Patch Changes
+
+- Updated dependencies [f31d016]
+  - @aave/client@4.0.0-next.16
+
+## 4.1.0-next.8
+
+### Patch Changes
+
+- Updated dependencies [402281c]
+  - @aave/client@4.0.0-next.15
+
 ## 4.1.0-next.7
 
 ### Patch Changes