+++ packages/client/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @aave/client
 
+## 4.0.0-next.16
+
+### Patch Changes
+
+- f31d016: **fix:** issue with decodeUserPositionId`.
+- Updated dependencies [f31d016]
+  - @aave/graphql@1.0.0-next.14
+
+## 4.0.0-next.15
+
+### Patch Changes
+
+- 402281c: **feat:** expose tx hash from `TransactionError`.
+- Updated dependencies [402281c]
+  - @aave/core@1.0.0-next.9
+
 ## 4.0.0-next.14
 
 ### Patch Changes